WIP: Getting http working
Have most of the endpoints working and removed. The error handling is a
major PITA though.
