+++ proxy/coco/Cargo.toml
@@ -14,6 +14,7 @@ serde = { version = "1.0", features = [ "derive" ] }
 syntect = "4.2"
 thiserror = "1.0"
 tokio = { version = "0.2", features = [ "dns", "macros", "time" ] }
+warp = { git = "https://github.com/radicle-dev/warp", branch = "openapi", features = [ "openapi" ], default-features = false }
 
 [dependencies.radicle-git-helpers]
 git = "https://github.com/radicle-dev/radicle-link.git"
