+++ proxy/http/src/control.rs
@@ -4,96 +4,59 @@ use serde::{Deserialize, Serialize};
 use warp::filters::BoxedFilter;
 use warp::{path, Filter, Rejection, Reply};
 
-use crate::http;
-use crate::registry;
+use crate::{Ctx, with_context, with_owner_guard};
 
 /// Combination of all control filters.
-pub fn filters<R>(ctx: http::Ctx<R>) -> BoxedFilter<(impl Reply,)>
-where
-    R: registry::Client + 'static,
+pub fn filters(ctx: Ctx) -> BoxedFilter<(impl Reply,)>
 {
     create_project_filter(ctx.clone())
         .or(nuke_coco_filter(ctx.clone()))
-        .or(nuke_registry_filter(ctx.clone()))
-        .or(register_user_filter(ctx))
         .boxed()
 }
 
 /// POST /create-project
-fn create_project_filter<R>(
-    ctx: http::Ctx<R>,
+fn create_project_filter(
+    ctx: Ctx,
 ) -> impl Filter<Extract = impl Reply, Error = Rejection> + Clone
-where
-    R: registry::Client + 'static,
 {
     path!("create-project")
-        .and(super::with_context(ctx.clone()))
-        .and(super::with_owner_guard(ctx))
+        .and(with_context(ctx.clone()))
+        .and(with_owner_guard(ctx))
         .and(warp::body::json())
         .and_then(handler::create_project)
 }
 
-/// POST /register-user
-fn register_user_filter<R>(
-    ctx: http::Ctx<R>,
-) -> impl Filter<Extract = impl Reply, Error = Rejection> + Clone
-where
-    R: registry::Client + 'static,
-{
-    path!("register-user")
-        .and(http::with_context(ctx))
-        .and(warp::body::json())
-        .and_then(handler::register_user)
-}
-
 /// GET /nuke/coco
-fn nuke_coco_filter<R>(
-    ctx: http::Ctx<R>,
+fn nuke_coco_filter(
+    ctx: Ctx,
 ) -> impl Filter<Extract = impl Reply, Error = Rejection> + Clone
-where
-    R: registry::Client + 'static,
 {
     path!("nuke" / "coco")
-        .and(super::with_context(ctx))
+        .and(with_context(ctx))
         .and_then(handler::nuke_coco)
-}
-
-/// GET /nuke/registry
-fn nuke_registry_filter<R>(
-    ctx: http::Ctx<R>,
-) -> impl Filter<Extract = impl Reply, Error = Rejection> + Clone
-where
-    R: registry::Client + 'static,
-{
-    path!("nuke" / "registry")
-        .and(http::with_context(ctx))
-        .and_then(handler::nuke_registry)
 }
 
 /// Control handlers for conversion between core domain and http request fulfilment.
 mod handler {
-    use radicle_registry_client::CryptoPair;
     use std::convert::TryFrom;
     use warp::http::StatusCode;
     use warp::{reply, Rejection, Reply};
 
     use librad::paths;
 
-    use crate::coco;
+    use coco;
+    use core::keystore;
+    use core::project;
+
     use crate::error::Error;
-    use crate::http;
-    use crate::keystore;
-    use crate::project;
-    use crate::registry;
+    use crate::Ctx;
 
     /// Create a project from the fixture repo.
-    pub async fn create_project<R>(
-        ctx: http::Ctx<R>,
+    pub async fn create_project(
+        ctx: Ctx,
         owner: coco::User,
         input: super::CreateInput,
     ) -> Result<impl Reply, Rejection>
-    where
-        R: Send + Sync,
     {
         let ctx = ctx.read().await;
 
@@ -123,38 +86,8 @@ mod handler {
         ))
     }
 
-    /// Register a user with another key
-    pub async fn register_user<R>(
-        ctx: http::Ctx<R>,
-        input: super::RegisterInput,
-    ) -> Result<impl Reply, Rejection>
-    where
-        R: registry::Client,
-    {
-        let ctx = ctx.read().await;
-
-        let fake_pair =
-            radicle_registry_client::ed25519::Pair::from_legacy_string(&input.handle, None);
-
-        log::info!(
-            "Registering user handle {} with public key {}",
-            &input.handle,
-            &fake_pair.public()
-        );
-
-        let handle = registry::Id::try_from(input.handle).map_err(Error::from)?;
-        ctx.registry
-            .register_user(&fake_pair, handle.clone(), None, input.transaction_fee)
-            .await
-            .expect("unable to register user");
-
-        Ok(reply::json(&true))
-    }
-
     /// Reset the coco state by creating a new temporary directory for the librad paths.
-    pub async fn nuke_coco<R>(ctx: http::Ctx<R>) -> Result<impl Reply, Rejection>
-    where
-        R: Send + Sync,
+    pub async fn nuke_coco(ctx: Ctx) -> Result<impl Reply, Rejection>
     {
         // TmpDir deletes the temporary directory once it DROPS.
         // This means our new directory goes missing, and future calls will fail.
@@ -181,18 +114,6 @@ mod handler {
         let mut ctx = ctx.write().await;
         ctx.peer_api = new_peer_api;
         ctx.keystore = new_keystore;
-
-        Ok(reply::json(&true))
-    }
-
-    /// Reset the Registry state by replacing the emulator in place.
-    pub async fn nuke_registry<R>(ctx: http::Ctx<R>) -> Result<impl Reply, Rejection>
-    where
-        R: registry::Client,
-    {
-        let (client, _) = radicle_registry_client::Client::new_emulator();
-        let mut ctx = ctx.write().await;
-        ctx.registry.reset(client);
 
         Ok(reply::json(&true))
     }
@@ -249,15 +170,6 @@ pub struct CreateInput {
     default_branch: String,
     /// Create and track fake peers
     fake_peers: Option<Vec<String>>,
-}
-/// Input for user registration.
-#[derive(Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RegisterInput {
-    /// Handle of the user.
-    handle: String,
-    /// User specified transaction fee.
-    transaction_fee: registry::Balance,
 }
 
 #[allow(clippy::unwrap_used)]
